Move proxy config into client
