+++ blaze-client/src/main/scala/org/http4s/client/blaze/BlazeClient.scala
@@ -74,11 +74,11 @@ object BlazeClient {
 
       config.proxy.lift(key) match {
         case Some(proxyConfig) =>
+          logger.debug(s"Proxying request to ${key} through ${proxyConfig.proxyHost}:${proxyConfig.proxyPort}")
           proxy(proxyConfig)
         case None =>
           manager.borrow(key).flatMap(loop(req))
       }
     }, onShutdown)
   }
-
 }
