+++ client/src/main/scala/org/http4s/client/ProxyConfig.scala
@@ -1,6 +1,5 @@
 package org.http4s
 package client
-package blaze
 
 import org.http4s.Uri.Authority
 import org.http4s.util.CaseInsensitiveString
