+++ client/src/main/scala/org/http4s/client/ProxySelector.scala
@@ -1,0 +1,4 @@
+package org.http4s
+package client
+
+object SystemPropertiesProxySelector
